+++ src/Rask.Ui/UiInput.cs
@@ -57,10 +57,8 @@ public sealed partial class UiInput<T> : Component, IFormControl<T>
     public Expression<Func<T>>? Bind { get; set; }
 
     /// <inheritdoc />
-    public Validate<T>? Validate { get; set; }
+    public Validator<T>? Validate { get; set; }
 
-    /// <inheritdoc />
-    public ValidateAsync<T>? ValidateAsync { get; set; }
 
     /// <inheritdoc />
     public Callback<T>? AfterBind { get; set; }
@@ -76,7 +74,6 @@ public sealed partial class UiInput<T> : Component, IFormControl<T>
             return Input
                 .Bind(bind)
                 .Validate(Validate)
-                .ValidateAsync(ValidateAsync)
                 .AfterBind(AfterBind)
                 .Type(Type)
                 .Placeholder(Placeholder ?? string.Empty)