+++ src/Rask.Ui/UiRadio.cs
@@ -56,10 +56,8 @@ public sealed partial class UiRadio : Component, IFormControl<bool>
     public Expression<Func<bool>>? Bind { get; set; }
 
     /// <inheritdoc />
-    public Validate<bool>? Validate { get; set; }
+    public Validator<bool>? Validate { get; set; }
 
-    /// <inheritdoc />
-    public ValidateAsync<bool>? ValidateAsync { get; set; }
 
     /// <inheritdoc />
     public Callback<bool>? AfterBind { get; set; }
@@ -76,7 +74,6 @@ public sealed partial class UiRadio : Component, IFormControl<bool>
             return Input
                 .Bind(bind)
                 .Validate(Validate)
-                .ValidateAsync(ValidateAsync)
                 .AfterBind(AfterBind)
                 .Type(InputType.Radio)
                 .Name(Group)