+++ src/Rask.Ui/UiSelect.cs
@@ -125,10 +125,8 @@ public sealed partial class UiSelect<T> : Component, IFormControl<T>
     public Expression<Func<T>>? Bind { get; set; }
 
     /// <inheritdoc />
-    public Validate<T>? Validate { get; set; }
+    public Validator<T>? Validate { get; set; }
 
-    /// <inheritdoc />
-    public ValidateAsync<T>? ValidateAsync { get; set; }
 
     /// <inheritdoc />
     public Callback<T>? AfterBind { get; set; }
@@ -160,7 +158,6 @@ public sealed partial class UiSelect<T> : Component, IFormControl<T>
             return Select
                 .Bind(bind)
                 .Validate(Validate)
-                .ValidateAsync(ValidateAsync)
                 .AfterBind(AfterBind)
                 .Aria(Aria(expanded: null))
                 .Disabled(Disabled == true)