+++ src/Rask.Ui/UiTextarea.cs
@@ -48,10 +48,8 @@ public sealed partial class UiTextarea<T> : Component, IFormControl<T>
     public Expression<Func<T>>? Bind { get; set; }
 
     /// <inheritdoc />
-    public Validate<T>? Validate { get; set; }
+    public Validator<T>? Validate { get; set; }
 
-    /// <inheritdoc />
-    public ValidateAsync<T>? ValidateAsync { get; set; }
 
     /// <inheritdoc />
     public Callback<T>? AfterBind { get; set; }
@@ -65,7 +63,6 @@ public sealed partial class UiTextarea<T> : Component, IFormControl<T>
             return Textarea
                 .Bind(bind)
                 .Validate(Validate)
-                .ValidateAsync(ValidateAsync)
                 .AfterBind(AfterBind)
                 .Placeholder(Placeholder ?? string.Empty)
                 .Rows(Rows ?? 3)